server: split sandboxes actions
